code/requirements.txt
torch
torchvision
pillow
numpy
pandas
tqdm
